Move java's JDAO into its own package.
Before this, JDAO shared a model ID with the JS JDAO and model IDs
should be unique.
