+++ src/foam/nanos/dig/DUGDAO.java
@@ -11,8 +11,8 @@ import foam.core.FObject;
 import foam.core.X;
 import foam.dao.AbstractSink;
 import foam.dao.DAO;
-import foam.dao.JDAO;
 import foam.dao.ProxyDAO;
+import foam.dao.java.JDAO;
 import foam.nanos.auth.User;
 import foam.nanos.logger.Logger;
 import foam.util.Auth;
