+++ src/foam/nanos/export/ExportService.java
@@ -7,9 +7,9 @@
 package foam.nanos.export;
 
 import foam.dao.DAO;
-import foam.dao.JDAO;
 import foam.dao.MapDAO;
 import foam.dao.ProxyDAO;
+import foam.dao.java.JDAO;
 import foam.nanos.NanoService;
 
 public class ExportService
