+++ src/foam/nanos/notification/email/services.jrl
@@ -1,3 +1,3 @@
-p({"class":"foam.nanos.boot.NSpec", "name":"emailTemplateDAO",                 "lazy":false,  "serve":true,  "serviceScript":"return new foam.dao.PMDAO(x, new foam.dao.SequenceNumberDAO(new foam.dao.JDAO(x, foam.nanos.notification.email.EmailTemplate.getOwnClassInfo(), \"emailTemplates\")));","client":"{\"of\":\"foam.nanos.notification.email.EmailTemplate\"}" })
+p({"class":"foam.nanos.boot.NSpec", "name":"emailTemplateDAO",                 "lazy":false,  "serve":true,  "serviceScript":"return new foam.dao.PMDAO(x, new foam.dao.SequenceNumberDAO(new foam.dao.java.JDAO(x, foam.nanos.notification.email.EmailTemplate.getOwnClassInfo(), \"emailTemplates\")));","client":"{\"of\":\"foam.nanos.notification.email.EmailTemplate\"}" })
 p({"class":"foam.nanos.boot.NSpec", "name":"emailMessageDAO",                    "lazy":false,  "serve":true,  "serviceScript":"return new foam.nanos.auth.CreatedAwareDAO.Builder(x).setDelegate( new foam.nanos.auth.CreatedByAwareDAO.Builder(x).setDelegate(new foam.dao.SequenceNumberDAO( new foam.dao.FixedSizeDAO.Builder(x).setDelegate(new foam.nanos.notification.email.SMTPEmailMessageDAO.Builder(x).setDelegate(new foam.dao.WriteOnlyJDAO(x, foam.nanos.notification.email.EmailMessage.getOwnClassInfo(), \"emails\")).build()).build())).build()).build();", "client":"{\"of\":\"foam.nanos.notification.email.EmailMessage\"}"})
 p({"class":"foam.nanos.boot.NSpec", "name":"email",                            "lazy":false,  "serve":true,  "authenticate": false, "boxClass":"foam.nanos.notification.email.EmailServiceSkeleton","client":"{ \"class\":\"foam.nanos.notification.email.ClientEmailService\", \"delegate\": { \"class\":\"foam.box.HTTPBox\", \"url\":\"service/email\" } }","serviceScript":"return new foam.nanos.notification.email.DAOEmailService.Builder(x).setDao((foam.dao.DAO)x.get(\"emailMessageDAO\")).build()"});
