+++ tools/classes.js
@@ -76,7 +76,7 @@ var classes = [
   'foam.box.SessionReplyBox',
   'foam.box.SessionClientBox',
   'foam.dao.DAO',
-  'foam.dao.JDAO',
+  'foam.dao.java.JDAO',
   'foam.dao.FilteredDAO',
   'foam.dao.Journal',
   'foam.dao.ProxyJournal',
